refactor(market): separate persistence from external API calls (#23)
- Extract MarketSnapshotPersistenceService to avoid @Transactional covering external I/O
- collectAndSave() calls collectors outside transaction, delegates DB write to persistence service
- Prevents self-invocation issue per CLAUDE.md transaction rules
