+++ src/test/java/com/solv/wefin/domain/market/service/MarketSnapshotServiceTest.java
@@ -2,7 +2,6 @@ package com.solv.wefin.domain.market.service;
 
 import com.solv.wefin.domain.market.collector.MarketDataCollector;
 import com.solv.wefin.domain.market.dto.CollectedMarketData;
-import com.solv.wefin.domain.market.entity.MarketSnapshot;
 import com.solv.wefin.domain.market.entity.MarketSnapshot.ChangeDirection;
 import com.solv.wefin.domain.market.entity.MarketSnapshot.MetricType;
 import com.solv.wefin.domain.market.entity.MarketSnapshot.Unit;
@@ -15,9 +14,9 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.math.BigDecimal;
 import java.util.List;
-import java.util.Optional;
 
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.*;
 
@@ -25,6 +24,9 @@ import static org.mockito.Mockito.*;
 class MarketSnapshotServiceTest {
 
     private MarketSnapshotService marketSnapshotService;
+
+    @Mock
+    private MarketSnapshotPersistenceService persistenceService;
 
     @Mock
     private MarketSnapshotRepository marketSnapshotRepository;
@@ -36,72 +38,39 @@ class MarketSnapshotServiceTest {
     private MarketDataCollector bokCollector;
 
     @Test
-    @DisplayName("수집 성공 — 신규 지표는 INSERT된다")
-    void collectAndSave_insert() {
+    @DisplayName("수집 성공 — persistenceService에 저장을 위임한다")
+    void collectAndSave_success() {
         // Given
         marketSnapshotService = new MarketSnapshotService(
-                List.of(yahooCollector), marketSnapshotRepository);
+                List.of(yahooCollector), persistenceService, marketSnapshotRepository);
 
         given(yahooCollector.collect()).willReturn(List.of(createNasdaqData()));
         given(yahooCollector.getSourceName()).willReturn("YahooFinance");
-        given(marketSnapshotRepository.findByMetricType(MetricType.NASDAQ))
-                .willReturn(Optional.empty());
 
         // When
         marketSnapshotService.collectAndSave();
 
         // Then
-        verify(marketSnapshotRepository).save(any(MarketSnapshot.class));
+        verify(persistenceService).saveSnapshots(anyList());
     }
 
     @Test
-    @DisplayName("수집 성공 — 기존 지표는 UPDATE된다")
-    void collectAndSave_update() {
-        // Given
-        marketSnapshotService = new MarketSnapshotService(
-                List.of(yahooCollector), marketSnapshotRepository);
-
-        MarketSnapshot existing = MarketSnapshot.builder()
-                .metricType(MetricType.NASDAQ)
-                .label("나스닥")
-                .value(new BigDecimal("20000"))
-                .changeRate(BigDecimal.ZERO)
-                .changeValue(BigDecimal.ZERO)
-                .unit(Unit.POINT)
-                .changeDirection(ChangeDirection.FLAT)
-                .build();
-
-        given(yahooCollector.collect()).willReturn(List.of(createNasdaqData()));
-        given(yahooCollector.getSourceName()).willReturn("YahooFinance");
-        given(marketSnapshotRepository.findByMetricType(MetricType.NASDAQ))
-                .willReturn(Optional.of(existing));
-
-        // When
-        marketSnapshotService.collectAndSave();
-
-        // Then — UPDATE이므로 save를 호출하지 않는다.
-        verify(marketSnapshotRepository, never()).save(any(MarketSnapshot.class));
-    }
-
-    @Test
-    @DisplayName("부분 실패 — 한 수집기 실패해도 다른 수집기는 정상 동작한다")
+    @DisplayName("부분 실패 — 한 수집기 실패해도 다른 수집기 결과는 저장된다")
     void collectAndSave_partialFailure() {
         // Given
         marketSnapshotService = new MarketSnapshotService(
-                List.of(yahooCollector, bokCollector), marketSnapshotRepository);
+                List.of(yahooCollector, bokCollector), persistenceService, marketSnapshotRepository);
 
         given(yahooCollector.collect()).willThrow(new RuntimeException("Yahoo 장애"));
         given(yahooCollector.getSourceName()).willReturn("YahooFinance");
         given(bokCollector.collect()).willReturn(List.of(createBaseRateData()));
         given(bokCollector.getSourceName()).willReturn("BOK");
-        given(marketSnapshotRepository.findByMetricType(MetricType.BASE_RATE))
-                .willReturn(Optional.empty());
 
         // When
         marketSnapshotService.collectAndSave();
 
-        // Then — BOK 데이터만 저장
-        verify(marketSnapshotRepository).save(any(MarketSnapshot.class));
+        // Then — BOK 데이터만 저장 위임
+        verify(persistenceService).saveSnapshots(anyList());
     }
 
     @Test
@@ -109,7 +78,7 @@ class MarketSnapshotServiceTest {
     void collectAndSave_allFailure() {
         // Given
         marketSnapshotService = new MarketSnapshotService(
-                List.of(yahooCollector, bokCollector), marketSnapshotRepository);
+                List.of(yahooCollector, bokCollector), persistenceService, marketSnapshotRepository);
 
         given(yahooCollector.collect()).willThrow(new RuntimeException("Yahoo 장애"));
         given(yahooCollector.getSourceName()).willReturn("YahooFinance");
@@ -120,7 +89,7 @@ class MarketSnapshotServiceTest {
         marketSnapshotService.collectAndSave();
 
         // Then
-        verify(marketSnapshotRepository, never()).save(any());
+        verify(persistenceService, never()).saveSnapshots(anyList());
     }
 
     @Test
@@ -128,7 +97,7 @@ class MarketSnapshotServiceTest {
     void collectAndSave_emptyResult() {
         // Given
         marketSnapshotService = new MarketSnapshotService(
-                List.of(yahooCollector), marketSnapshotRepository);
+                List.of(yahooCollector), persistenceService, marketSnapshotRepository);
 
         given(yahooCollector.collect()).willReturn(List.of());
         given(yahooCollector.getSourceName()).willReturn("YahooFinance");
@@ -137,8 +106,7 @@ class MarketSnapshotServiceTest {
         marketSnapshotService.collectAndSave();
 
         // Then
-        verify(marketSnapshotRepository, never()).save(any());
-        verify(marketSnapshotRepository, never()).findByMetricType(any());
+        verify(persistenceService, never()).saveSnapshots(anyList());
     }
 
     private CollectedMarketData createNasdaqData() {
